// src/helpers/proto/plugnmeet_breakout_room_pb.ts
export class CreateBreakoutRoomsReq extends Message<CreateBreakoutRoomsReq> {
  /**
   * @generated from field: string room_id = 1;
   */
  roomId = '';

  /**
   * @generated from field: string requested_user_id = 2;
   */
  requestedUserId = '';

  /**
   * @generated from field: uint64 duration = 3;
   */
  duration = protoInt64.zero;

  /**
   * @generated from field: optional string welcome_msg = 4;
   */
  welcomeMsg?: string;

  /**
   * @generated from field: repeated plugnmeet.BreakoutRoom rooms = 5;
   */
  rooms: BreakoutRoom[] = [];

  constructor(data?: PartialMessage<CreateBreakoutRoomsReq>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = 'plugnmeet.CreateBreakoutRoomsReq';
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: 'room_id', kind: 'scalar', T: 9 /* ScalarType.STRING */ },
    {
      no: 2,
      name: 'requested_user_id',
      kind: 'scalar',
      T: 9 /* ScalarType.STRING */,
    },
    { no: 3, name: 'duration', kind: 'scalar', T: 4 /* ScalarType.UINT64 */ },
    {
      no: 4,
      name: 'welcome_msg',
      kind: 'scalar',
      T: 9 /* ScalarType.STRING */,
      opt: true,
    },
    { no: 5, name: 'rooms', kind: 'message', T: BreakoutRoom, repeated: true },
  ]);

  static fromBinary(
    bytes: Uint8Array,
    options?: Partial<BinaryReadOptions>,
  ): CreateBreakoutRoomsReq {
    return new CreateBreakoutRoomsReq().fromBinary(bytes, options);
  }

  static fromJson(
    jsonValue: JsonValue,
    options?: Partial<JsonReadOptions>,
  ): CreateBreakoutRoomsReq {
    return new CreateBreakoutRoomsReq().fromJson(jsonValue, options);
  }

  static fromJsonString(
    jsonString: string,
    options?: Partial<JsonReadOptions>,
  ): CreateBreakoutRoomsReq {
    return new CreateBreakoutRoomsReq().fromJsonString(jsonString, options);
  }

  static equals(
    a:
      | CreateBreakoutRoomsReq
      | PlainMessage<CreateBreakoutRoomsReq>
      | undefined,
    b:
      | CreateBreakoutRoomsReq
      | PlainMessage<CreateBreakoutRoomsReq>
      | undefined,
  ): boolean {
    return proto3.util.equals(CreateBreakoutRoomsReq, a, b);
  }
}

/**
 * @generated from message plugnmeet.BreakoutRoom
 */
export class BreakoutRoom extends Message<BreakoutRoom> {
  /**
   * @generated from field: string id = 1;
   */
  id = '';

  /**
   * @generated from field: string title = 2;
   */
  title = '';

  /**
   * @generated from field: uint64 duration = 3;
   */
  duration = protoInt64.zero;

  /**
   * @generated from field: bool started = 4;
   */
  started = false;

  /**
   * @generated from field: uint64 created = 5;
   */
  created = protoInt64.zero;

  /**
   * @generated from field: repeated plugnmeet.BreakoutRoomUser users = 6;
   */
  users: BreakoutRoomUser[] = [];

  constructor(data?: PartialMessage<BreakoutRoom>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = 'plugnmeet.BreakoutRoom';
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: 'id', kind: 'scalar', T: 9 /* ScalarType.STRING */ },
    { no: 2, name: 'title', kind: 'scalar', T: 9 /* ScalarType.STRING */ },
    { no: 3, name: 'duration', kind: 'scalar', T: 4 /* ScalarType.UINT64 */ },
    { no: 4, name: 'started', kind: 'scalar', T: 8 /* ScalarType.BOOL */ },
    { no: 5, name: 'created', kind: 'scalar', T: 4 /* ScalarType.UINT64 */ },
    {
      no: 6,
      name: 'users',
      kind: 'message',
      T: BreakoutRoomUser,
      repeated: true,
    },
  ]);

  static fromBinary(
    bytes: Uint8Array,
    options?: Partial<BinaryReadOptions>,
  ): BreakoutRoom {
    return new BreakoutRoom().fromBinary(bytes, options);
  }

  static fromJson(
    jsonValue: JsonValue,
    options?: Partial<JsonReadOptions>,
  ): BreakoutRoom {
    return new BreakoutRoom().fromJson(jsonValue, options);
  }

  static fromJsonString(
    jsonString: string,
    options?: Partial<JsonReadOptions>,
  ): BreakoutRoom {
    return new BreakoutRoom().fromJsonString(jsonString, options);
  }

  static equals(
    a: BreakoutRoom | PlainMessage<BreakoutRoom> | undefined,
    b: BreakoutRoom | PlainMessage<BreakoutRoom> | undefined,
  ): boolean {
    return proto3.util.equals(BreakoutRoom, a, b);
  }
}

/**
 * @generated from message plugnmeet.BreakoutRoomUser
 */
export class BreakoutRoomUser extends Message<BreakoutRoomUser> {
  /**
   * @generated from field: string id = 1;
   */
  id = '';

  /**
   * @generated from field: string name = 2;
   */
  name = '';

  /**
   * @generated from field: bool joined = 3;
   */
  joined = false;

  constructor(data?: PartialMessage<BreakoutRoomUser>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = 'plugnmeet.BreakoutRoomUser';
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: 'id', kind: 'scalar', T: 9 /* ScalarType.STRING */ },
    { no: 2, name: 'name', kind: 'scalar', T: 9 /* ScalarType.STRING */ },
    { no: 3, name: 'joined', kind: 'scalar', T: 8 /* ScalarType.BOOL */ },
  ]);

  static fromBinary(
    bytes: Uint8Array,
    options?: Partial<BinaryReadOptions>,
  ): BreakoutRoomUser {
    return new BreakoutRoomUser().fromBinary(bytes, options);
  }

  static fromJson(
    jsonValue: JsonValue,
    options?: Partial<JsonReadOptions>,
  ): BreakoutRoomUser {
    return new BreakoutRoomUser().fromJson(jsonValue, options);
  }

  static fromJsonString(
    jsonString: string,
    options?: Partial<JsonReadOptions>,
  ): BreakoutRoomUser {
    return new BreakoutRoomUser().fromJsonString(jsonString, options);
  }

  static equals(
    a: BreakoutRoomUser | PlainMessage<BreakoutRoomUser> | undefined,
    b: BreakoutRoomUser | PlainMessage<BreakoutRoomUser> | undefined,
  ): boolean {
    return proto3.util.equals(BreakoutRoomUser, a, b);
  }
}